6_django_module/django3_drf_auth/mysite/blog/serializers.py
from django import forms
from rest_framework import serializers

from blog.models import Post


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['title', 'content', 'demo_content']
        # fields = '__all__'
